MeshBuilder

+++ libGPU/src/main/java/com/monstrous/graphics/VertexAttribute.java
@@ -21,13 +21,13 @@ import com.monstrous.webgpu.WGPUVertexFormat;
 public class VertexAttribute {
 
     public static class Usage {
-        static public final long POSITION = 1;
-        static public final long COLOR = 2;
-        static public final long TEXTURE_COORDINATE = 4;
-        static public final long NORMAL= 8;
-        static public final long TANGENT = 16;
-        static public final long BITANGENT = 32;
-        static public final long GENERIC = 64;
+        static public final int POSITION = 1;
+        static public final int COLOR = 2;
+        static public final int TEXTURE_COORDINATE = 4;
+        static public final int NORMAL= 8;
+        static public final int TANGENT = 16;
+        static public final int BITANGENT = 32;
+        static public final int GENERIC = 64;
     }
